refactor(common): break data-app code/types import cycle + harden validation [GLITCH-559]
Move ApiGetAppCodeResponse, ImportAppCodeRequestBody, ApiImportAppCodeResponse from
types.ts to code.ts so the only dependency edge is code.ts → types.ts. Add a guard
for non-object file entries and extend the test suite with absolute-path and
non-object file-entry cases.
